+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Make `Iterable.groupBy` and `Iterable.indexed` return instances of `MapEntry` for easier interoperability with `Map`.
 * Fix various edge cases with `Range` iterables.
 * Deprecated `naturalComparator` in favor of `naturalCompare` and `naturalComparable<T>`.
+* Add `List` extensions `takeTo`, `skipTo`, `takeLast`, `takeLastTo`, `skipLast`, and `skipLastTo`.
 
 ## 3.8.0
 * Add `String` extensions `take`, `takeTo`, `skip`, `skipTo`, `takeLast`, `takeLastTo`, `skipLast`, and `skipLastTo`.